+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - DecaLog now collects extended metrics for: PHP, WordPress core, database, WordPress multisite and plugins / themes using [WordPress DecaLog SDK](https://github.com/Pierre-Lannoy/wp-decalog-sdk).
 - DecaLog now collects minimalist metrics for Wordfence.
 - DecaLog now collects extended traces for WordPress core and plugins / themes using [WordPress DecaLog SDK](https://github.com/Pierre-Lannoy/wp-decalog-sdk).
+- New logger to send events to Datadog.
 - New logger to send metrics to a Prometheus instance (via Pushgateway).
 - New logger to publish metrics as endpoint for Prometheus scrapping.
 - New logger to send metrics to InfluxDB Cloud or to an on-premise InfluxDB 2 instance or cluster.